Fixed styling issues with djangocms-admin-style
djangocms-admin-style expects two `field-box`es side by side to be
of equal width. This commit fixes the styles so by default the pair
of values is fluid and stretches to the full width of the
container.

Also it changes the event handlers on delete to be delegated so
it's now possible to remove pairs that were just created.